Cont. work on search / car list function
+ some minor css tweaking before the test release
