+++ cars/static/cars/css/spesific_carStyle.css
@@ -1,22 +1,11 @@
 #car_image_container {
-  border-top-left-radius: 50px;
-  border-top-right-radius: 50px;
-  background-color: #5C6B63;
+  border-top-left-radius: 5px;
+  border-top-right-radius: 5px;
+  background-color: #143840;
   padding-bottom: 20px; }
-
-body {
-  background-color: #143840; }
 
 p {
   font-size: x-large; }
-
-.navbar-static-top {
-  background-color: #5C6B63; }
-  .navbar-static-top a {
-    color: white; }
-
-.navbar-static-top a {
-  color: white; }
 
 #car_container {
   padding-top: 40px; }
@@ -40,24 +29,25 @@ p {
   padding-bottom: 20px; }
 
 .btn {
-  background-color: #143840;
+  background-color: #A69E89;
   color: white; }
 
 .btn:hover {
-  background-color: #143840;
-  color: #D96523; }
+  background-color: #E0C297;
+  color: black; }
 
 #reservation_box {
   background-color: #143840;
-  border-top-left-radius: 50px;
-  border-top-right-radius: 50px;
+  border-top-left-radius: 5px;
+  border-top-right-radius: 5px;
   margin-bottom: 20px;
   margin-top: 20px; }
 
 #car_info_container {
-  background-color: #5C6B63;
-  border-bottom-left-radius: 50px;
-  border-bottom-right-radius: 50px; }
+  background-color: #143840;
+  border-bottom-left-radius: 10px;
+  border-bottom-right-radius: 10px;
+  margin-bottom: 30px; }
 
 #car_info_section {
   margin: 20px; }
