+++ cars/templates/cars/spesific_car.html
@@ -47,9 +47,7 @@
 
     <!-- Container END -->
     </div>
-</section>
 
-<section id="car_info_section">
 
     <div class="container" id="car_info_container">
         <div class="row">
