+++ static/baseStyle/css/baseStyle.css
@@ -16,8 +16,10 @@ body {
 #push {
   height: 100px; }
 
-.navbar-static-top {
+.navbar {
   background-color: #143840; }
+  .navbar a {
+    color: white; }
 
 .footer {
   background-color: dimgrey;
