+++ static/baseStyle/css/baseStyle.sass
@@ -27,8 +27,15 @@ body
   height: 100px
 
 
-.navbar-static-top
+.navbar
   background-color: $color1
+
+  a
+    color: white
+
+
+
+
 
 .footer
   background-color: dimgrey
