+++ static/booking/css/booking_form_style.css
@@ -6,4 +6,13 @@ body {
   #form_container label {
     font-size: large; }
 
+p {
+  font-size: larger; }
+
+#car_info_table .table td {
+  width: 80%;
+  font-size: 150%; }
+#car_info_table .table #total {
+  border-bottom: 2px solid #fff; }
+
 /*# sourceMappingURL=booking_form_style.css.map */
