+++ static/booking/css/booking_form_style.sass
@@ -17,3 +17,17 @@ body
 
   label
     font-size: large
+
+
+
+p
+  font-size: larger
+
+#car_info_table
+  .table
+    td
+      width: 80%
+      font-size: 150%
+
+    #total
+      border-bottom: 2px solid #fff
