+++ static/booking/css/receipt_style.css
@@ -1,0 +1,4 @@
+.table td {
+  font-size: 150%; }
+
+/*# sourceMappingURL=receipt_style.css.map */
