+++ static/booking/css/receipt_style.sass
@@ -1,0 +1,10 @@
+$color1: #143840
+$color2: #5C6B63
+$color3: #A69E89
+$color4: #E0C297
+$color5: #D96523
+
+.table
+  td
+    font-size: 150%
+    //border: 1px solid black
