+++ static/cars/css/car_list_style.css
@@ -1,0 +1,37 @@
+.car_tab {
+  border: solid;
+  border-color: #143840;
+  background-color: #143840;
+  border-radius: 5px;
+  padding-top: 20px;
+  margin-top: 20px; }
+  .car_tab .btn {
+    background-color: #E0C297;
+    color: #5C6B63;
+    vertical-align: middle; }
+  .car_tab .btn:hover, .car_tab .btn:focus {
+    color: black;
+    background-color: #D96523; }
+
+#price_p {
+  text-decoration: underline;
+  font-size: xx-large;
+  padding-top: 20px; }
+
+.car_info p {
+  font-size: x-large; }
+
+.car_price_info p {
+  font-size: xx-large; }
+.car_price_info div {
+  display: inline-block;
+  vertical-align: bottom;
+  float: none;
+  margin: -2px; }
+
+.car_images a img {
+  width: 360px;
+  height: 240px;
+  border: none; }
+
+/*# sourceMappingURL=car_list_style.css.map */
