+++ static/cars/css/car_list_style.sass
@@ -1,0 +1,53 @@
+$color1: #143840
+$color2: #5C6B63
+$color3: #A69E89
+$color4: #E0C297
+$color5: #D96523
+
+
+.car_tab
+  border: solid
+  border-color: $color1
+  background-color: $color1
+  border-radius: 5px
+  padding-top: 20px
+  margin-top: 20px
+
+
+  .btn
+    background-color: $color4
+    color: $color2
+    vertical-align: middle
+
+  .btn:hover, .btn:focus
+    color: black
+    background-color: $color5
+
+
+
+#price_p
+  text-decoration: underline
+  font-size: xx-large
+  padding-top: 20px
+
+.car_info
+  p
+    font-size: x-large
+
+
+.car_price_info
+  p
+    font-size: xx-large
+  div
+    display: inline-block
+    vertical-align: bottom
+    float: none
+    margin: -2px
+
+
+.car_images
+  a
+    img
+      width: 360px
+      height: 240px
+      border: none
