+++ static/cars/css/spesific_carStyle.sass
@@ -12,17 +12,15 @@ $color4: #E0C297
 $color5: #D96523
 
 #car_image_container
-  border-top-left-radius: 50px
-  border-top-right-radius: 50px
-  background-color: $color2
-  padding-bottom: 20px
-
-
-body
+  border-top-left-radius: 5px
+  border-top-right-radius: 5px
   background-color: $color1
+  padding-bottom: 20px
 
 p
   font-size: x-large
+
+
 
 #car_container
   padding-top: 40px
@@ -46,21 +44,30 @@ p
   padding-top: 20px
   padding-bottom: 20px
 
+.btn
+  background-color: $color3
+  color: white
+
+.btn:hover
+  background-color: $color4
+  color: black
+
 #reservation_box
   background-color: $color1
   //border-radius: 10px
-  border-top-left-radius: 50px
-  border-top-right-radius: 50px
+  border-top-left-radius: 5px
+  border-top-right-radius: 5px
   margin-bottom: 20px
   margin-top: 20px
 
 
 
 #car_info_container
-  background-color: $color2
+  background-color: $color1
   //border-radius: 5px
-  border-bottom-left-radius: 50px
-  border-bottom-right-radius: 50px
+  border-bottom-left-radius: 10px
+  border-bottom-right-radius: 10px
+  margin-bottom: 30px
 
 #car_info_section
   margin: 20px
